# api/openapi.yaml
openapi: 3.1.0
info:
  title: 'COGIT Construction API'
  description: 'API for construction scheduling and intervention management'
  version: 1.0.0
servers:
  -
    url: /
    description: ''
paths:
  /api/register:
    post:
      operationId: api_register
      tags:
        - Authentication
      summary: 'Register a new user'
      description: 'Creates a new user account'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
      responses:
        '201':
          description: 'User created successfully'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      firstName:
                        type: string
                      lastName:
                        type: string
        '400':
          description: 'Invalid input'
        '409':
          description: 'Email already registered'
  /api/login:
    post:
      operationId: api_login
      tags:
        - Authentication
      summary: 'Login to get JWT token'
      description: 'Authenticates user and returns JWT access token'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  format: email
                  description: 'User email address'
                password:
                  type: string
      responses:
        '200':
          description: 'Login successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: 'JWT access token (valid for 1 hour)'
        '401':
          description: 'Invalid credentials'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT token obtained from /api/login'
security:
  - bearerAuth: []
...